+++ src/daaf/evalexps/evaluation.py
@@ -11,11 +11,11 @@ from typing import Any, Iterator, Mapping, Optional, Set
 import numpy as np
 from numpy.typing import DTypeLike
 from daaf import core
+from daaf.core import GeneratesEpisode
 from daaf.learning.opt import schedules
 from daaf.learning.tabular import policies, policyeval
 
 from daaf import constants, expconfig, options, task, utils
-from daaf.evalexps import methods
 
 
 def run_fn(experiment_run: expconfig.ExperimentRun):
@@ -109,7 +109,7 @@ def evaluate_policy(
     num_episodes: int,
     algorithm: str,
     learnign_args: expconfig.LearningArgs,
-    generate_steps_fn: core.GeneratesEpisode,
+    generate_steps_fn: GeneratesEpisode,
 ) -> Iterator[policyeval.PolicyEvalSnapshot]:
     """
     Runs policy evaluation with given algorithm, env, and policy spec.
@@ -118,18 +118,20 @@ def evaluate_policy(
     eval_fn: Optional[Iterator[policyeval.PolicyEvalSnapshot]] = None
     if algorithm == constants.ONE_STEP_TD:
         if daaf_config.traj_mapping_method == constants.DAAF_TRAJECTORY_MAPPER:
-            eval_fn = methods.onpolicy_one_step_td_state_values_only_aggregate_updates(
-                policy=policy,
-                environment=env_spec.environment,
-                num_episodes=num_episodes,
-                lrs=schedules.LearningRateSchedule(
-                    initial_learning_rate=learnign_args.learning_rate,
-                    schedule=task.constant_learning_rate,
-                ),
-                gamma=learnign_args.discount_factor,
-                state_id_fn=env_spec.discretizer.state,
-                initial_values=initial_state_values,
-                generate_episode=generate_steps_fn,
+            eval_fn = (
+                policyeval.onpolicy_one_step_td_state_values_only_aggregate_updates(
+                    policy=policy,
+                    environment=env_spec.environment,
+                    num_episodes=num_episodes,
+                    lrs=schedules.LearningRateSchedule(
+                        initial_learning_rate=learnign_args.learning_rate,
+                        schedule=task.constant_learning_rate,
+                    ),
+                    gamma=learnign_args.discount_factor,
+                    state_id_fn=env_spec.discretizer.state,
+                    initial_values=initial_state_values,
+                    generate_episode=generate_steps_fn,
+                )
             )
         else:
             eval_fn = policyeval.onpolicy_one_step_td_state_values(
@@ -153,7 +155,7 @@ def evaluate_policy(
             daaf_config.traj_mapping_method
             == constants.DAAF_NSTEP_TD_UPDATE_MARK_MAPPER
         ):
-            eval_fn = methods.nstep_td_state_values_on_aggregate_start_steps(
+            eval_fn = policyeval.nstep_td_state_values_on_aggregate_start_steps(
                 policy=policy,
                 environment=env_spec.environment,
                 num_episodes=num_episodes,
@@ -185,7 +187,7 @@ def evaluate_policy(
 
     elif algorithm == constants.FIRST_VISIT_MONTE_CARLO:
         if daaf_config.traj_mapping_method == constants.DAAF_TRAJECTORY_MAPPER:
-            eval_fn = methods.onpolicy_first_visit_monte_carlo_state_values_only_aggregate_updates(
+            eval_fn = policyeval.onpolicy_first_visit_monte_carlo_state_values_only_aggregate_updates(
                 policy=policy,
                 environment=env_spec.environment,
                 num_episodes=num_episodes,
